LayoutTests/platform/chromium/accessibility/main-element.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<main id="main">main</main>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the HTML5 main element correctly maps to AXLandmarkMain.");

    if (window.accessibilityController) {

          document.getElementById("body").focus();
          var body = accessibilityController.focusedElement;
          var main = accessibilityController.accessibleElementById("main");
          shouldBe("main.role", "'AXRole: AXLandmarkMain'");

    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
