LayoutTests/platform/chromium/accessibility/main-element-expected.txt

main
This tests that the HTML5 main element correctly maps to AXLandmarkMain.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS main.role is 'AXRole: AXLandmarkMain'
PASS successfullyParsed is true

TEST COMPLETE
